[chore]: improve the code
