+++ packages/shared/lib/enum/index.ts
@@ -60,7 +60,8 @@ export enum StorageKey {
   Token = 'TOKEN',
   DMS_Project_Catch = 'DMS_Project_Catch',
   USER_UID = 'DMS_USER_UID',
-  SHOW_COMPANY_NOTICE = 'SHOW_COMPANY_NOTICE'
+  SHOW_COMPANY_NOTICE = 'SHOW_COMPANY_NOTICE',
+  DMS_CB_CHANNEL = 'DMS_CB_CHANNEL'
 }
 
 /**
